+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.0.6
+- Allow customize column header background colors
+- Allow developer to update task limit for column via board provider
+- Bug fixes
+
 ## 0.0.5 
 - Enhance Kanban theme with customizable border properties for columns and task cards and **default** to borderless
 - Set background color for confirmation and task form dialogs